+++ .web_package/hooks/dev/2_demo_symlink.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# 
+# @file
+# Copy relevent files from parent into demo
+#
+test -e "$7/demo/jquery.loft_labels.js" && rm "$7/demo/jquery.loft_labels.js"
+cd "$7/demo/" && ln -s "$7/jquery.loft_labels.js" .